+++ NEWS.md
@@ -1,3 +1,10 @@
+# Version 0.9.2
+
+- GENCODE transcript tracs:
+   - Pull in more up-to-date MANE select transcripts from NCBI
+   - Pull in RefSeq Select identifiers (grch37)
+   - Use exact version numbers for transcripts/MANE select etc.
+
 # Version 0.9.1
 
 - Use up-to-date gene symbols from NCBI in GENCODE tracks, ignore outdated Entrez